requirements.txt
pandas==1.5.3
numpy==1.24.3
plotly-express==0.4.1
streamlit==1.28.1
image==1.5.33
folium==0.14.0
streamlit-folium==0.13.0
haversine==2.8.0
scipy==1.10.1
statsmodels==0.13.5